Updated from altool to NotaryTool
Added -A Auto run managedsoftwareupdate immediately after install. This
                really should be used only with DEP/ADM enrollments.
Removed -m since its deprecated and not in use anymore

Updated the notarize script to use NotaryTool instead of altool.

Altool will be deprecated fall 2023.

Updated the Readme to have a guide for notarytool and some information regarding this.

Cleaned up the Readme so its more correct